[Dubbo] move the classes of model from config to rpc-api for reusing
